+++ rust/operator-binary/src/main.rs
@@ -1,8 +1,9 @@
-mod authentication;
 mod config;
 mod controller;
 mod operations;
 mod product_logging;
+mod reporting_task;
+mod security;
 
 use clap::{crate_description, crate_version, Parser};
 use futures::stream::StreamExt;